% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myfunction}
\alias{myfunction}
\title{Square of a vector}
\usage{
myfunction(x)
}
\arguments{
\item{x}{}
}
\value{
Returns square of a vector
}
\description{
Square of a vector
}
\examples{
x<- 1:30 myfunction(x)
}
