% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{A constructor function for t tests}
\usage{
myttest(x, y, alpha = 0.05)
}
\arguments{
\item{x}{vector of data}

\item{y}{vector of data}

\item{alpha}{signifigance level between 0 and 1}
}
\description{
A constructor function for t tests
}
\examples{
 set.seed(21);x <-rnorm(30,5,2);
 set.seed(23); y<- rnorm(40, 3,2);
 alpha <- 0.05
 obj <- myttest(x = x, y = y, alpha = 0.05)
 class(obj)
 print(obj)
}
