% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print the Rttest Object}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{Rttest object}

\item{...}{needed for print to work}
}
\description{
Print the Rttest Object
}
